+++ Changelog.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased] Valkyrie Version 0.2 Changelog
 
+### Added
+
+- Warning screen system to notify of potential frictions between Valkyrie and other mods
+- Added warning for using OptiFine
+
 ### Changed
 
 - Changed the upper limit for the `Zoom Multiplier` to 10